backend/requirements.txt
Django
psycopg2-binary
djangorestframework
python-decouple
django-cors-headers
Pillow
boto3
python-dateutil
celery
redis
django-filter
drf-spectacular
